chore: separate faucet into component
